Move python 2 enum code into separate file
